requirements.txt

gymnasium[mujoco]
pygame>=2.1.3
pre-commit
stable-baselines3
sbx-rl
torchsummary
torchviz
moviepy
scikit-learn
deap
wandb>=0.21.0
ipykernel>=6.30.1
rl-zoo3>=2.7.0
tensorboard>=2.20.0
jax[cuda]
